Update loss_functions.py
I had a wrong understanding of the function. I added comments for the class differentiation to make this clearer and added a fix for log(0) not being defined by setting the loss for log(0) to -log(eps). Finally, to ensure that the loss function is continuous, the minimal value of the computed loss and the capped loss is returned. This ensures that -log(x) with x < eps will not have -log(x) > -log(0).